Use windowTitle instead of getModel in _cleaunup_panel
If someone tries to setModel on not exported or somehow broken device, getModel returns empty string and TaurusDevicePanels is broken, so after trying to close the widget, it cannot find and remove it. Then, after reopening (even with restarted device) the panel, it finds broken widget again and doesn't try to open the new one.

`Name` is also used for setting window title and in this case it's always the same.